Hide regulars warning on cyclic chain first shift ONLY if no regulars
Warnings should be shown on first shifts too if there are regulars for
the target location. This means that the target is not a new location,
but there was a new chain made for it (Maybe an old client).

(varoitukset pois toistuvan ketjun ekalla vuorolla ainoastaan jos
kohteella ei ole omasiistijöitä; muuten, varoitukset normaalisti)