fixed week calendar hiding shifts when updating in certain conditions
basically if 2 employees had the same repeating shift, and you'd update a shift that wasn't the foremention repeating shift, which also didn't have the 2 employees as a pair, the other employees slot would update in a way which only showed the shifts that have both employees as a pair.
to fix this we just need to remove the employee IDs from the result of tv_arr that weren't defined in the update call.